Remove sed dependency
sed is used to configure input headers during build. The
only reason for this is so that the arch directory can be
determined at make time.

Since for now, we only care about x86_64, we check in alltypes.h
and syscall.h after they are generated for that architecture.
This removes the dependency on sed.